Task 1

+++ Roguelike/Actors/IDamageable.cs
@@ -1,6 +1,9 @@
 ﻿namespace Roguelike.Actors;
 
+/// <summary>
+///     Данный интерфейс отвечает за возможность актора получать урон
+/// </summary>
 public interface IDamageable
 {
-    public void TakeDamage(float damage);
+    public void TakeDamage(int damage);
 }
